Rename Task3 to rock_paper_scissors-Task1
rock , paper , scissors is an game which we are since a long period.
It is a simple and understandable python code to play rock , paper and scissors.
It is user friendly.
It takes user name and invites you to play game .
Computer asks the user to choose rock or paper or scissors . As user should pick first. Further computer picks it's choice randomly. Next it shows the result of winner and displays the score .
You can play many rounds.