Rename Task2 to Password_Generator-Task2
In this time security matters alot . Everybody wants to be secured and safe secure ( money , personal data etc). 
Today information is wealth. So all should be aware of cyber crimes. All should keep their passwords strong.
This is an python code of a Password Generator, which gives a strong password to user . The complexity of password is defined by the user .
This gives user a strong and random password.